Create IntegrityCheck.scala
When I initially started planning BigBrainSecurity.scala I was overwhelmed with trying to understand recursion and I was more concerned with experimenting w/ data structures than I was with getting the program done. The updated version of BigBrainSecurity which is a subclass of IntegrityCheck.scala will be much more practical. Scala has a steep learning curve, but once you learn it it's incredibly powerful. I have a feeling that someday Scala will be my programming language of choice over python and that I'll be able to work in Scala more efficiently than I do in python. 

The secret to programming is planning. I drew a program diagram last night so it should be a lot easier to write this program now. Assuming I can figure out how to make the map function or flatmap work. And that the Scala compiler stops being difficult.